requirements.txt
sai_mujoco
sai_rl
torch
imageio
huggingface_hub
hf_xet
stable-baselines3
wandb
tqdm
numpy==2.1.3
jax
flax
distrax
tensorflow==2.19.0
scikit-learn
PyYAML